Fixed "Unable to create ShellObject" bug
This bug occurs when a location is unable to be opened in BetterExplorer
(probably because it's not a valid directory path), so when this occurs
during startup, an error message shows and it will load the default tab
in that tab's place.

Also adds in code and other such things in preparation for my Tabs
update.