Localize Tab Manager, fix Folder Tools tab issues
Localized all strings in Tab Manager and related windows. However, some
layout issues still need to be fixed in cases where the string is longer
than the button. Also fixes issues related to the Folder Tools tab
appearing when it shouldn't, and the Archive Tools tab not appearing at
all.