Added Greek, fix Search Tools translation issues
Added Greek translation to BetterExplorer, and then dealt with errors
involving it for an hour. Note that the translation is not perfect.
Changed translation handling for criteria dialogs on Search Tools tab,
so they state "Set <value> filter" (or equivalent in their language).
Some translations may not be 100% correct as the result.