Begin FTP/WebDAV/Dropbox support, add /norestore
Made a lot of the UI for the upcoming FTP/WebDAV server support, and for
online services such as Dropbox and Box.net. Also added support for the
/norestore switch, which will cause that instance of Better Explorer to
not restore the previous session. (Will also be good if one of the tabs
is causing issues.) Also made small UI and layout fixes, and added or
refined some KeyTips on the File tab.