Finish FTP file system support (no UI yet though)
Added code for connecting the AlexFTPS client to the unified network
file system. Unfortunately, with the AddServer window not yet working, I
cannot test to see if FTP support is good. Full support for FTPS is not
yet added, but will use the same connections to the file system that FTP
uses.