Add FTPS, WebDAV, and Dropbox file system support
Added Sharpbox for WebDAV and Dropbox support (as well as other
services), and added a window for oAuth requests for Dropbox (and
eventually SkyDrive). Also added FTPS support, and greatly simplified
code for adding servers (as well as adding support for editing
already-added servers). Other modifications and improvements for the
network features. Also re-organized some code in MainWindow.xaml.cs.