Add more icons, minor functionality changes
Add icons for the new commands, and as well adds icons to existing
commands that were missing icons. Also changes to make sure the Easy
Access menu is disabled when it cannot be used. A checkmark is supposed
to appear next to check boxes in the QAT customize window, although it
seems to not be working for me.