Add UI for Mount ISO as Drive
Created a new Virtual Disks Tools tab which is displayed when an ISO,
VHD, or BIN file is selected, and has a "Mount as Drive" command, taking
advantage of the ImDisk driver. Note that if you don't have the driver
installed, an error will occur but the program should still continue to
function.