Added more ISO-related options, reorganized code
Created a Mount ISO dialog that will allow a user to set additional
options for mounting an ISO file. Also added a "Write to Disk" command
that will open Windows Disk Image Burner. Added an "Unmount" option on
the Drive Tools tab that will unmount a virtual drive. Note that with
this addition, Better Explorer takes a small moment longer to become
responsive when you select a drive in the Computer folder. Also added
better error handling for if ImDisk is not installed, with an option for
the user to download it if it is not installed. Hid the "Use Vista-style
list view" check in the Options. This option can still be changed using
the appropriate Registry key. Also did some code reorganization.